+++ CHANGELOG.md
@@ -21,6 +21,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Added `created_at` to the mail subscription objects in `/api/v1/users/user-preferences` API. remp/respekt#301
 - Fixed newsletter list seeder. remp/remp#1391
 - Added `FrontendPresenter` for identification of presenters available to public. remp/remp#1395
+- Added `update()` (to updated `updated_at`) methods to `BatchesRepository`, `JobsRepository`, `LogsRepository`, `SourceTemplatesRepository`. remp/remp#1397
 
 ## Archive